Update Trace plugin to have Tracer class's functionality, remove Tracer.
