+++ src/zelos/ext/platforms/linux/loader.py
@@ -52,7 +52,7 @@ class LinuxMode:
     def handle_exception(self, p, e):
         # TODO(kzsnow): This goes in core?
         try:
-            self.z.trace.bb(
+            self.z.plugins.trace.bb(
                 self.z.last_instruction, self.z.last_instruction_size
             )
         except Exception:
@@ -79,13 +79,13 @@ class LinuxMode:
             if self._attempt_to_handle_syscall():
                 return  # linear execution after syscall (interrupt style)
 
-        self.z.trace.bb()
+        self.z.plugins.trace.bb()
         p.threads.fail_current_thread(fail_reason=f"Exception {e}")
         self.z.processes.handles.close_all(self.z.current_process.pid)
 
     def _attempt_to_handle_syscall(self):
-        if self.z.verbose:
-            self.z.trace.bb(
+        if self.z.plugins.trace.verbose:
+            self.z.plugins.trace.bb(
                 self.z.last_instruction,
                 self.z.last_instruction_size,
                 full_trace=False,
