+++ src/zelos/ext/plugins/trace.py
@@ -15,9 +15,12 @@
 # <http://www.gnu.org/licenses/>.
 # ======================================================================
 
+import logging
+
 import capstone.arm_const as cs_arm
 import capstone.x86_const as cs_x86
 
+from termcolor import colored
 from unicorn import UC_ERR_READ_UNMAPPED, UcError
 
 from zelos import HookType, IPlugin
@@ -25,12 +28,30 @@ from zelos import HookType, IPlugin
 
 class Trace(IPlugin):
     def __init__(self, z):
-        super().__init__(z)
+        self.zelos = z
+
+        self.logger = logging.getLogger(__name__)
+
+        self.current_return_address = 0
+        self.current_function_name = "???"
+        self.current_api_module = "???"
+
+        # Comments (to be dumped)
+        # We should save a lot of space by moving this to a grouping by
+        # thread, rather than keeping the thread with each comment.
+        # This will break compatibility with systems like Doppler.
+        self.comments = []
+        self.MAX_INDENTS = 40
+        self.threads_to_print = set()
+        if z.config.tracethread != "":
+            self.threads_to_print.add(z.config.tracethread)
 
         self.verbosity = z.config.verbosity
         self.verbose = False
         self.set_verbose(self.verbosity > 0)
         self.fasttrace = True if z.config.fasttrace > 0 else False
+        self.trace_on = ""
+        self.trace_off = ""
 
         self.last_instruction = None
         self.last_instruction_size = None
@@ -38,6 +59,40 @@ class Trace(IPlugin):
 
         if self.verbose:
             self.set_hook_granularity(HookType.EXEC.INST)
+
+        if self.state.arch in ["x86", "x86_64"]:
+            self.comment_generator = x86CommentGenerator(
+                self.zelos, self.modules
+            )
+        elif self.state.arch == "arm":
+            self.comment_generator = ArmCommentGenerator(
+                self.zelos, self.modules
+            )
+        else:
+            self.comment_generator = EmptyCommentGenerator()
+
+    @property
+    def cs(self):
+        return self.zelos.internal_engine.cs
+
+    @property
+    def modules(self):
+        return self.zelos.internal_engine.modules
+
+    @property
+    def main_module(self):
+        return self.zelos.internal_engine.main_module
+
+    @property
+    def state(self):
+        return self.zelos.internal_engine.state
+
+    @property
+    def functions_called(self):
+        return self.comment_generator.functions_called
+
+    def get_region(self, addr):
+        return self.zelos.internal_engine.memory.get_region(addr)
 
     def set_hook_granularity(self, granularity: HookType.EXEC):
         """
@@ -80,9 +135,7 @@ class Trace(IPlugin):
             return
 
         self.zelos.thread.total_blocks_executed += 1
-        rev_modules = (
-            self.zelos.internal_engine.modules.reverse_module_functions
-        )
+        rev_modules = self.modules.reverse_module_functions
         if (
             self.zelos.thread.total_blocks_executed % 1000 == 0
             and address not in rev_modules
@@ -92,7 +145,7 @@ class Trace(IPlugin):
 
         if self.verbose:
             if self.should_print_last_instruction:
-                self.zelos.internal_engine.trace.bb(
+                self.zelos.internal_engine.plugins.trace.bb(
                     self.last_instruction,
                     self.last_instruction_size,
                     full_trace=False,
@@ -194,6 +247,197 @@ class Trace(IPlugin):
         else:
             self.set_hook_granularity(HookType.EXEC.BLOCK)
 
+    def bb(self, address=None, size=20, full_trace=False):
+        if not self.should_print_thread():
+            return
+        """ Prints instructions starting at the given address up to
+        'size' bytes away. If no address is given, prints starting at
+        the current address"""
+        if address is None:
+            address = self.zelos.regs.getIP()
+        try:
+            code = self.zelos.internal_engine.emu.mem_read(address, size)
+            insns = [insn for insn in self.cs.disasm(code, address)]
+            if len(insns) == 0:
+                return
+            # For full trace, we'll just print the first instruction,
+            # and then all the registers
+            if full_trace:
+                self.ins(insns[0])
+                self.regs()
+            else:
+                for insn in insns:
+                    self.ins(insn)
+        except UcError as e:
+            if e.errno == UC_ERR_READ_UNMAPPED:
+                print("Unable to read instruction at address %x" % address)
+            else:
+                raise e
+
+    def regs(self):
+        """ Prints registers at the current address"""
+        s = ""
+        reg_list = self.zelos.internal_engine.emu.imp_regs
+        for reg in reg_list:
+            s += " ".join(
+                [f"{reg}={self.zelos.internal_engine.emu.get_reg(reg):x}"]
+            )
+            s += "\n"
+        print(s)
+
+    # There are issues when this is used with autohooks. Need to see how
+    # we can include this in the future. without endless indents
+    def indent(self):
+        self.zelos.thread._callstack_indent_count += 1
+
+    def unindent(self):
+        self.zelos.thread._callstack_indent_count -= 1
+
+    def set_current_return_address(self, addr):
+        self.current_return_address = addr
+
+    def set_current_function_name(self, name):
+        self.current_function_name = name
+
+    def set_current_api_module(self, name):
+        self.current_api_module = name
+
+    def print(self, category, s, thread=None, addr_str=None):
+        if thread is None:
+            thread = self.zelos.thread.name
+        if addr_str is None:
+            addr_str = f"{self.zelos.regs.getIP():08x}"
+        thread_str = colored(f"[{thread}]", "magenta")
+        category_str = colored(f"[{category}]", "red")
+        addr_str_str = colored(f"[{addr_str}]", "white", attrs=["bold"])
+        print(f"{thread_str} {category_str} {addr_str_str} {s}")
+
+    def log_api(self, args, isNative=False):
+        self.api(args, isNative)
+
+    def log_api_dbg(self, args):
+        self.api_dbg(args)
+
+    # Prints the thread, return address and api string
+    def api(self, args, isNative=False):
+        """ Prints an API that was called"""
+        if not self.should_print_thread():
+            return
+        return_address = self.current_return_address
+        indent_count = self.zelos.thread._callstack_indent_count
+        try:
+            caller_module = (
+                self.get_region(return_address).module_name.split(".")[0]
+                + "_____"
+            )[:8]
+        except Exception:
+            caller_module = "________"
+        native_s = ""
+        if isNative:
+            native_s = "[Native] "
+        if indent_count == -1:
+            indent_count = 0
+        args = "".join([i if ord(i) < 128 else "." for i in args])
+        s = (
+            "  " * min(indent_count, self.MAX_INDENTS)
+            + f"{native_s}{self.current_api_module}!{args}"
+        )
+        self.print("API", s, addr_str=f"{caller_module}:{return_address:08x}")
+
+    # Prints the thread, return address and api string
+    def api_dbg(self, args, isNative=False):
+        """ Prints an API that was called"""
+        if not self.should_print_thread():
+            return
+        return_address = self.current_return_address
+        indent_count = self.zelos.thread._callstack_indent_count
+        caller_module = (
+            self.get_region(return_address).module_name.split(".")[0] + "_____"
+        )[:8]
+        if indent_count == -1:
+            indent_count = 0
+        args = "".join([i if ord(i) < 128 else "." for i in args])
+        s = "  " * min(indent_count, self.MAX_INDENTS) + colored(
+            args, "white", attrs=["bold"]
+        )
+        self.print("API", s, addr_str=f"{caller_module}:{return_address:08x}")
+
+    def ins(self, insn):
+        """ Prints the thread, address and instruction string """
+        if not self.should_print_thread():
+            return
+        sep = ""
+        if insn.address == self.zelos.regs.getIP():
+            sep = "*"
+        address = insn.address
+        ins_string = self._get_insn_string(insn)
+        if address in self.main_module.exported_functions:
+            function_name = self.main_module.exported_functions[address]
+            s = colored(f"<{function_name}>", "white", attrs=["bold"])
+            self.print("INS", s, addr_str=f"{sep}{address:08x}")
+        self.print("INS", ins_string, addr_str=f"{sep}{address:08x}")
+
+    def should_print_thread(self, t=None):
+        """
+        Decides whether log statements should be printed for the given
+        thread
+        """
+        # If the thread is known to be benign, let's ignore it. We can
+        # add a config to print these if we need it.
+        if t is None:
+            t = self.zelos.thread
+            if t is None:
+                return False
+
+        if t.benign_code is True:
+            return False
+        # The user can choose to print threads or to focus only on
+        # specific threads.
+        if len(self.threads_to_print) == 0:
+            return True
+        return t.name in self.threads_to_print
+
+    def _get_insn_string(self, insn):
+        """ Gets the string to be printed for an instruction."""
+        cmt = ""
+        try:
+            cmt = self.comment_generator.get_comment(insn)
+        except Exception as e:
+            self.logger.notice(
+                f"Issue printing {insn.mnemonic} instruction comment: {e}"
+            )
+
+        result = ""
+        insn_str = "{0}\t{1}".format(insn.mnemonic, insn.op_str)
+        if len(cmt) > 0:
+            padding = ""
+            padSize = 60 - len(insn_str)
+            if padSize < 0:
+                padSize = 1
+            for y in range(0, padSize):
+                padding += " "
+            result += (
+                insn_str
+                + " "
+                + padding
+                + colored(" ; " + cmt, "grey", attrs=["bold"])
+            )
+            # Log comment for the dump
+            self.comments.append(
+                Comment(insn.address, self.zelos.thread.id, cmt)
+            )
+        else:
+            result += insn_str
+
+        return result
+
+
+class Comment:
+    def __init__(self, address, thread_id, text):
+        self.address = address
+        self.thread_id = thread_id
+        self.text = text
+
 
 class EmptyCommentGenerator:
     def get_comment(self, insn):
@@ -201,10 +445,9 @@ class EmptyCommentGenerator:
 
 
 class ArmCommentGenerator:
-    def __init__(self, z, tracer, modules):
+    def __init__(self, zelos, modules):
         self.functions_called = {}
-        self._z = z
-        self.tracer = tracer
+        self.zelos = zelos
         self._modules = modules
 
     def get_comment(self, insn):
@@ -246,13 +489,12 @@ class ArmCommentGenerator:
         """
         Returns the syscall name
         """
+        sm = self.zelos.internal_engine.zos.syscall_manager
         if insn.insn_name() == "svc" and insn.operands[0].imm != 0:
             syscall_num = insn.operands[0].imm - 0x900000
         else:
-            syscall_num = self._z.zos.syscall_manager.get_syscall_number()
-        syscall_name = self._z.zos.syscall_manager.find_syscall_name_by_number(
-            syscall_num
-        )
+            syscall_num = sm.get_syscall_number()
+        syscall_name = sm.find_syscall_name_by_number(syscall_num)
         return f"{syscall_name}"
 
     def _dst_comment(self, insn):
@@ -289,8 +531,12 @@ class ArmCommentGenerator:
         Returns a comment on branch to label.
         """
         src_val = self._get_reg_or_mem_val(insn, insn.operands[0])
-        if src_val in self._z.main_module.exported_functions:
-            func_name = self._z.main_module.exported_functions[src_val]
+        if (
+            src_val
+            in self.zelos.internal_engine.main_module.exported_functions
+        ):
+            main_module = self.zelos.internal_engine.main_module
+            func_name = main_module.exported_functions[src_val]
             return f"<{func_name:s}> (0x{src_val:x})"
         return f"<0x{src_val:x}>"
 
@@ -300,32 +546,38 @@ class ArmCommentGenerator:
         the memory value at the location specified
         """
         if x.type == cs_arm.ARM_OP_REG:
-            return self.tracer.emu.get_reg(insn.reg_name(x.value.reg))
+            return self.zelos.internal_engine.emu.get_reg(
+                insn.reg_name(x.value.reg)
+            )
         elif x.type == cs_arm.ARM_OP_IMM:
             return x.imm
         else:
             base_val = (
                 0
                 if x.mem.base == 0
-                else self.tracer.emu.get_reg(insn.reg_name(x.mem.base))
+                else self.zelos.internal_engine.emu.get_reg(
+                    insn.reg_name(x.mem.base)
+                )
             )
             shift_val = (
                 0
                 if x.mem.index == 0
-                else self.tracer.emu.get_reg(insn.reg_name(x.mem.index))
+                else self.zelos.internal_engine.emu.get_reg(
+                    insn.reg_name(x.mem.index)
+                )
                 * x.mem.scale
             )
             if is_dst:
                 return base_val + shift_val + x.value.mem.disp
             else:
-                return self.tracer.memory.read_int(
+                return self.zelos.memory.read_int(
                     base_val + shift_val + x.value.mem.disp
                 )
 
 
 class x86CommentGenerator:
-    def __init__(self, tracer, modules):
-        self.tracer = tracer
+    def __init__(self, zelos, modules):
+        self.zelos = zelos
         self._modules = modules
         self.functions_called = {}
 
@@ -333,7 +585,7 @@ class x86CommentGenerator:
         """Returns a string representing the data pointed to by 'ptr' if 'ptr'
         is a valid pointer. Otherwise, reutrns an empty string."""
         try:
-            pointer_data = self.tracer.memory.read_int(ptr)
+            pointer_data = self.zelos.memory.read_int(ptr)
         except UcError as e:
             if e.errno == UC_ERR_READ_UNMAPPED:
                 return ""
@@ -341,7 +593,7 @@ class x86CommentGenerator:
 
         s = ""
         try:
-            s = self.tracer.memory.read_string(ptr, 8)
+            s = self.zelos.memory.read_string(ptr, 8)
         except UcError as e:
             if e.errno != UC_ERR_READ_UNMAPPED:
                 raise e
@@ -371,7 +623,7 @@ class x86CommentGenerator:
         # Only used when looking at the current instruction.
         # op = operands[0]
         # target = op.value.imm
-        target = self.tracer.emu.getIP()
+        target = self.zelos.internal_engine.emu.getIP()
         self.functions_called[target] = True
         cmt = insn.mnemonic + "(0x{0:x}) ".format(target)
         if target in self._modules.reverse_module_functions:
@@ -427,24 +679,30 @@ class x86CommentGenerator:
         memory value at the location specified
         """
         if x.type == cs_x86.X86_OP_REG:
-            return self.tracer.emu.get_reg(insn.reg_name(x.value.reg))
+            return self.zelos.internal_engine.emu.get_reg(
+                insn.reg_name(x.value.reg)
+            )
         elif x.type == cs_x86.X86_OP_IMM:
             return x.imm
         else:
             base_val = (
                 0
                 if x.mem.base == 0
-                else self.tracer.emu.get_reg(insn.reg_name(x.mem.base))
+                else self.zelos.internal_engine.emu.get_reg(
+                    insn.reg_name(x.mem.base)
+                )
             )
             shift_val = (
                 0
                 if x.mem.index == 0
-                else self.tracer.emu.get_reg(insn.reg_name(x.mem.index))
+                else self.zelos.internal_engine.emu.get_reg(
+                    insn.reg_name(x.mem.index)
+                )
                 * x.mem.scale
             )
             if is_dst:
                 return base_val + shift_val + x.value.mem.disp
             else:
-                return self.tracer.memory.read_int(
+                return self.zelos.memory.read_int(
                     base_val + shift_val + x.value.mem.disp, sz=x.size
                 )
