+++ tests/test_linux_x64.py
@@ -27,7 +27,7 @@ DATA_DIR = path.join(path.dirname(path.abspath(__file__)), "data")
 class ZelosTest(unittest.TestCase):
     def test_static_elf_unpatched(self):
         z = Zelos(path.join(DATA_DIR, "static_elf_x64_helloworld"))
-        z.internal_engine.trace.threads_to_print.add("none")
+        z.internal_engine.plugins.trace.threads_to_print.add("none")
         z.internal_engine.start(timeout=3)
 
         self.assertEqual(
@@ -36,7 +36,7 @@ class ZelosTest(unittest.TestCase):
 
     def test_static_elf(self):
         z = Zelos(path.join(DATA_DIR, "static_elf_x64_helloworld"))
-        z.internal_engine.trace.threads_to_print.add("none")
+        z.internal_engine.plugins.trace.threads_to_print.add("none")
         z.internal_engine.start(timeout=3)
 
         self.assertEqual(
@@ -45,7 +45,7 @@ class ZelosTest(unittest.TestCase):
 
     def test_dynamic_elf_unpatched(self):
         z = Zelos(path.join(DATA_DIR, "dynamic_elf_x64_helloworld"))
-        z.internal_engine.trace.threads_to_print.add("none")
+        z.internal_engine.plugins.trace.threads_to_print.add("none")
         z.internal_engine.start(timeout=3)
 
         self.assertEqual(
@@ -54,7 +54,7 @@ class ZelosTest(unittest.TestCase):
 
     def test_dynamic_elf(self):
         z = Zelos(path.join(DATA_DIR, "dynamic_elf_x64_helloworld"))
-        z.internal_engine.trace.threads_to_print.add("none")
+        z.internal_engine.plugins.trace.threads_to_print.add("none")
         z.internal_engine.start(timeout=3)
 
         self.assertEqual(
